+++ Changelog.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 - Fixed: Method 'SetStartDateOnCardAsync' was renamed to 'SetStartDateAndDueDateOnCardAsync' [COMPILE TIME BREAKING CHANGE]
+- Fixed: Smart-events did not do a proper internal await so could technically be delayed
 
 ## 1.1.0 (8th of Feb. 2023)
 - Added: Webhook System (See video how to get going [here](https://youtu.be/A3_B-SLBm_0))